functions/download_lyrics/requirements.txt
requests==2.25.1
pyjwt==2.1.0